Only allow increase of C++ standard via environment variable
Summary:
Tweak to the part of the CMake configuration that allows the version of the C++ standard to be modified via the environment variable `CMAKE_CXX_STANDARD`.
The modified implementation only allows the value to be changed if the specified standard is newer than our default, currently C++17.

Test Plan: Configured the project with the environment variable set to 14 and 20 and obtained the expected behaviour: no change, standard bumped to C++20.

Reviewers: jback, kreps, abudinen

Reviewed By: abudinen

Tags: #evtgen

Differential Revision: https://phab.hepforge.org/D106