Drop redundant EVTGEN_CPP11 preprocessor switch
Summary:
The EVTGEN_CPP11 preprocessor switch was introduced when compiler support for C++11 was variable.
Now we only support C++17 or later standards.
As such, this can be dropped.

This is related with [!16](https://gitlab.cern.ch/evtgen/evtgen/-/merge_requests/16).

Test Plan: Have run the full set of tests and the results are identical.

Reviewers: jback, kreps, abudinen

Reviewed By: abudinen

Subscribers: #evtgen_admins

Tags: #evtgen

Differential Revision: https://phab.hepforge.org/D117